Registration and login

+++ classes/Ijdb/Controllers/Joke.php
@@ -2,16 +2,19 @@
 
 namespace Ijdb\Controllers;
 use Core\DatabaseTable;
+use Core\Authentication;
 
 class Joke
 {
 	private $jokesTable;
 	private $authorsTable;
+	private $authentication;
 	
-	public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable)
+	public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication)
 	{
 		$this->jokesTable = $jokesTable;
 		$this->authorsTable = $authorsTable;
+		$this->authentication = $authentication;
 	}
 
 	public function list()
@@ -28,6 +31,7 @@ class Joke
 				'timestamp' => $joke['timestamp'],
 				'name' => $author['name'],
 				'email' => $author['email'],
+				'author_id' => $author['id']
 			];
 		}
 
@@ -35,18 +39,23 @@ class Joke
 
 		$totalJokes = $this->jokesTable->total();
 
+		$user = $this->authentication->getUser();
+
 		return [
 			'title' => $title,
 			'template' => 'jokes',
 			'variables' => [
 				'jokes' => $jokes,
-				'totalJokes' => $totalJokes
+				'totalJokes' => $totalJokes,
+				'user_id' => $user['id'] ?? null
 			]
 		];
 	}
 
 	public function form()
 	{
+		$user = $this->authentication->getUser();
+
 		if (isset($_GET['id'])) {
 			$joke = $this->jokesTable->findById($_GET['id']);
 			$title = 'Edit joke';
@@ -59,14 +68,23 @@ class Joke
 			'template' => 'joke_form',
 			'variables' => [
 				'joke' => $joke ?? null,
+				'user_id' => $user['id'] ?? null
 			]
 		];
 	}
 
 	public function saveForm()
 	{
+		$user = $this->authentication->getUser();
+
+		if (isset($_GET['id'])) {
+			$joke = $this->jokesTable->findById($_GET['id']);
+
+			if ($user['id'] !== $joke['author_id']) return;
+		}
+
 		$joke = $_POST['joke'];
-		$joke['author_id'] = 1;
+		$joke['author_id'] = $user['id'];
 
 		$this->jokesTable->save($joke);
 
@@ -75,6 +93,12 @@ class Joke
 
 	public function delete()
 	{
+		$user = $this->authentication->getUser();
+
+		$joke = $this->jokesTable->findById($_POST['id']);
+
+		if ($user['id'] !== $joke['author_id']) return;
+		
 		$this->jokesTable->delete($_POST['id']);
 
 		header('Location: /joke/list');
